refactor(grpc): move config
