+++ pkg/cmd/cmd_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/alexfalkowski/go-health/pkg/checker"
 	"github.com/alexfalkowski/go-health/pkg/server"
 	"github.com/alexfalkowski/go-service/pkg/cmd"
-	"github.com/alexfalkowski/go-service/pkg/config"
 	"github.com/alexfalkowski/go-service/pkg/health"
 	healthGRPC "github.com/alexfalkowski/go-service/pkg/health/transport/grpc"
 	healthHTTP "github.com/alexfalkowski/go-service/pkg/health/transport/http"
@@ -28,9 +27,7 @@ func TestInvalidHTTP(t *testing.T) {
 		os.Setenv("DATABASE_URL", "postgres://test:test@localhost:5432/test?sslmode=disable")
 
 		Convey("When I try to create a server", func() {
-			opts := []fx.Option{
-				logger.Module, pkgHTTP.Module, grpc.Module, config.Module,
-			}
+			opts := []fx.Option{logger.Module, pkgHTTP.Module, grpc.Module}
 
 			err := cmd.RunServer([]string{}, 10*time.Second, opts)
 
@@ -55,7 +52,7 @@ func TestInvalidGRPC(t *testing.T) {
 
 		Convey("When I try to create a server", func() {
 			opts := []fx.Option{
-				logger.Module, pkgHTTP.Module, grpc.Module, config.Module,
+				logger.Module, pkgHTTP.Module, grpc.Module,
 				health.Module, fx.Provide(registrations),
 				fx.Provide(httpObserver), fx.Provide(grpcObserver),
 			}
@@ -74,7 +71,7 @@ func TestInvalidGRPC(t *testing.T) {
 	})
 }
 
-func registrations(cfg *config.Config, logger *zap.Logger) health.Registrations {
+func registrations(logger *zap.Logger) health.Registrations {
 	hc := checker.NewHTTPCheckerWithClient("https://google.com", 1*time.Second, pkgHTTP.NewClient(logger))
 	hr := server.NewRegistration("http", 5*time.Second, hc)
