--- pkg/config/module.go
@@ -1,8 +1,0 @@
-package config
-
-import (
-	"go.uber.org/fx"
-)
-
-// Module for fx.
-var Module = fx.Provide(NewConfig)
